+++ Formular/src/org/formular/core/Operation.java
@@ -12,7 +12,7 @@ public abstract class Operation implements Serializable {
 	 */
 	private static final long serialVersionUID = 1L;
 
-	public abstract int result();
+	public abstract float result();
 
 	public abstract Collection <CardElement> inputDescriptions();
